+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Added custom difftool support by [Ivan Gilchrist](https://github.com/gilly3)
 * Updated Simplified Chinese translation by [Zhi](https://github.com/yfwz100)
 * Fixes for Brackets 0.44/1.0 by [Marcel Gerber](https://github.com/MarcelGerber)
+* Added revert change button to the gutter widgets.
 
 ## 0.13.30
 * New Dark UI styling by [Larz](https://github.com/larz0)